+++ packages/fluekit/src/FlexProps.ts
@@ -57,7 +57,7 @@ export type FlexBoxProps = {
   // 是否换行
   wrap?: boolean;
   // 间距
-  gap?: number;
+  gap?: number | string;
   // 是否填充可用空间
   expanded?: boolean;
   as?: string;